Reworked Workout page. Improved sets by adding timers, removed +/- buttons, added icon near set title, shrunk sets that are not selected, remade most of the logic and made it smoother by moving to the next set/exercise on set/field completion, removed completion checkbox from exercise list, and much more

+++ src/components/common/QuickGoalLog/QuickGoalLog.tsx
@@ -70,7 +70,7 @@ const QuickGoalLog: React.FC<LogGoalProps> = ({goalData, close, closeMenu}) => {
                 icon: goalData.icon,
                 timestamp: todayDate,
                 data: {
-                    value: goalData.type === 'number' || goalData.type === 'target' ? parseInt(inputValue) : goalData.type === 'yes-no' ? inputValue : inputValue,
+                    value: goalData.type === 'number' || goalData.type === 'target' ? parseFloat(inputValue) : goalData.type === 'yes-no' ? inputValue : inputValue,
                     time,
                     description,
                     name,
